#2810 - Move fact-linking stuff to a separate module
- Moved the stuff over to a separate module
- By default no longer include in the final application
