+++ inception/inception-kb-fact-linking/src/main/java/de/tudarmstadt/ukp/inception/kb/factlinking/feature/SubjectObjectFeatureEditor.properties
