+++ inception/inception-kb-fact-linking/src/main/resources/META-INF/asciidoc/user-guide/images/factExtraction5.png
