+++ inception/inception-kb-fact-linking/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+de.tudarmstadt.ukp.inception.ui.kb.config.FactLinkingAutoConfiguration
