+++ inception/inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/config/KnowledgeBaseServiceUIAutoConfiguration.java
@@ -22,7 +22,6 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
@@ -36,7 +35,6 @@ import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBaseProperties;
 import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBaseServiceAutoConfiguration;
 import de.tudarmstadt.ukp.inception.ui.kb.KnowledgeBasePageMenuItem;
 import de.tudarmstadt.ukp.inception.ui.kb.feature.ConceptFeatureSupport;
-import de.tudarmstadt.ukp.inception.ui.kb.initializers.FactLayerInitializer;
 import de.tudarmstadt.ukp.inception.ui.kb.initializers.NamedEntityIdentifierFeatureInitializer;
 import de.tudarmstadt.ukp.inception.ui.kb.project.KnowledgeBaseProjectSettingsPanelFactory;
 import de.tudarmstadt.ukp.inception.ui.kb.project.ProjectKnowledgeBaseMenuItem;
@@ -171,14 +169,5 @@ public class KnowledgeBaseServiceUIAutoConfiguration
     public ProjectKnowledgeBaseMenuItem projectKnowledgeBaseMenuItem()
     {
         return new ProjectKnowledgeBaseMenuItem();
-    }
-
-    @ConditionalOnProperty(prefix = "fact-layer", name = "enabled", havingValue = "true", matchIfMissing = false)
-    @Bean
-    @Autowired
-    public FactLayerInitializer factLayerInitializer(
-            AnnotationSchemaService aAnnotationSchemaService)
-    {
-        return new FactLayerInitializer(aAnnotationSchemaService);
     }
 }
